Don't cmake if build files have already been generated

+++ build_c.sh
@@ -1,5 +1,13 @@
-rm -Rf lib/build
+if [[ $1 == clean ]]; then
+    rm -Rf lib/build
+    exit 0
+fi
+
 mkdir -p lib/build
 cd lib/build
-cmake ..
+
+if [ ! -f "Makefile" ]; then
+    cmake ..
+fi
+
 make
